+++ odmf/db/__init__.py
@@ -2,5 +2,7 @@
 # -*- coding:utf-8 -*-
 
 from .base import Base, Session, engine, session_scope
-from .dbobjects import *
-from .dataset import *
+from .job import Job
+from .site import Site, Log, Datasource, Installation
+from .dbobjects import Person, Project, Image
+from .dataset import Dataset, DatasetGroup, DatasetItemGetter, Timeseries, TransformedTimeseries, Record, Quality, ValueType
